changed the RunwayParameters from ints to doubles to reflect the class diagram

+++ src/main/java/runway/Model/RunwayParameters.java
@@ -2,44 +2,44 @@ package runway.Model;
 
 public class RunwayParameters {
 
-    private int tora, toda, asda, lda;
+    private double tora, toda, asda, lda;
 
-    public RunwayParameters(int tora, int toda, int asda, int lda) {
+    public RunwayParameters(double tora, double toda, double asda, double lda) {
         this.tora = tora;
         this.toda = toda;
         this.asda = asda;
         this.lda = lda;
     }
 
-    public int getLda() {
+    public double getLda() {
         return lda;
     }
 
-    public void setLda(int lda) {
+    public void setLda(double lda) {
         this.lda = lda;
     }
 
-    public int getAsda() {
+    public double getAsda() {
         return asda;
     }
 
-    public void setAsda(int asda) {
+    public void setAsda(double asda) {
         this.asda = asda;
     }
 
-    public int getToda() {
+    public double getToda() {
         return toda;
     }
 
-    public void setToda(int toda) {
+    public void setToda(double toda) {
         this.toda = toda;
     }
 
-    public int getTora() {
+    public double getTora() {
         return tora;
     }
 
-    public void setTora(int tora) {
+    public void setTora(double tora) {
         this.tora = tora;
     }
 }
